fix FF/REW stability Improved FF/REW application stability
FF/REW Stability changes:

- Added a 3 stage FF/REW operation.

1. Pause Pipeline. (new API. Does same as STOP but does not reset PTS/DTS
Trimm Offset)
2. Seek Operation.
3. Start Pipeline.

- Added seeking of key frame after a FF/REW operation on/after a
specified time during seek. DashDownloader is unable to retrieve data at
specified point in time. It works on Dash Chunk resolution, thus seeked
content may be in the middle of chunk. Feeding SM Player data before
seeked position is... confusing SM Player.
To prevent it, when on OnSeekData() is called after a seek, offset
information is collected (absolute seek point) for Audio and Video.
Once BOTH offsets are collected, submit task is resumed.
In submit task, Audio/Video seeking is performed.
EOS/ConfigPackets are not ignored by seek - passed down for processing.
Audio Seeking is dependant on Video Seeking. For this very reason, both,
Audio and Video OnSeekData need to be recieved before starting Submit
Task.
When Audio reaches its seek position, it will wait for Video to do same,
as Video Seek may increase Audio Seek position due to search for a key
frame, not just specified time. Once Video Seek completes, audio seek
will be free to complete.

- Added init segment / non init segment differentiation
to HandleFailedDownload. In order not to introduce a separate
HandleFailedDownlaod for init/non init segments, a bool flag (defaulting
to false) has been added to DownloadSegment (imho, faster then doing
such selection in code based on ifs). This flag is passed down to
CreateDownloadTask and is a part of DownloadRequest. DashDownloader in
DownloadAsnc() will append this flag to exception thrown on download
error. This flag is retrieved by IsInitSegment() in DashClient and
passed to HandleFailedDownload().
This prevents two scenarios:

1. Call Stop() when waiting for INIT segment would result in Demux
throwing an exception resulting in misleading information passed back to
user.

2. Prevents a case where one pipeline dies while the other keeps on
going (If stop would not be called)

- Added OnStreamError notification from DashClient in case of failed
segments

FIXES:

- PlayerView.xmal.cs:
1. Changed PlayerView.xmal.cs popup display. Using Device.StartTimer() with
BeginInvokeOnMainThread() when displaying error popups, caused popup to
loose focus at random.
2. Added a block in PlayerView.xmal.cs preventing OnStreamEnd() handling
when error popup is displayed. This prevents closure of player while popup
is visible (resulting in loss of focus and inability to close popup)
3. Added PlayerController.Stop() when StreamError is recieved.

- SMPlayer.cs
When calling Stop in uninitialized state, player state is changed from
uninitialized to Stopping within a lock.
This prevents a possible race condition if player is performing SetStreamConfig.
During SetStreamConfig (case where it is done in Uninitialized State), State is set to
Ready and SubmitTask is started. If Stop gets called during that phase it would be silently
ignored.