fix Exception in DisposeEncryptedPackets Silent exception at SM Player StartUp
ResetPacketQueue was passing NULL queues (at initial startup) to
cleaning task. Task was not expecting null queues which resulted in
silent exception.

Observable at debug execution
Not seen ar runtime (non debug) execution

Expended DisposeEncryptedPacket() to expect any of the queues to be
null.
Cleaning task WILL NOT be started if both queues are null. If just one
is null (not really an expected scenario), clearing will be started.
With changes to Dispose method, it is now safe to clear null queues.